Updated UI to be friendlier.
Added several prompts and messages to improve the user experience. The
model of the program is essentially complete (might go back and factor a
Tracker class out). It still does not deal with certain Unicode
characters very well, but it can handle most. Next goal is to work on an
actual GUI.